refactor(v2): retarget EncoderTask onto tempo::EncoderInput
The concrete Arduino impl is renamed to rotary_encoder_input.h. on_start
replaces the first-poll arming flag — scheduler already fires it once
before any tick.

+++ include/v2/events.h
@@ -41,7 +41,7 @@ namespace pocketpd {
      * `delta` is the signed tick count since the previous sample.
      */
     struct EncoderEvent {
-        int16_t delta = 0;
+        int delta = 0;
     };
 
     using Event = tempo::Events<PdReadyEvent, ButtonEvent, EncoderEvent>;
